pyproject.toml
[project]
name = "nemotron-ts-posttrain"
version = "0.1.0"
description = "Post-training and evaluation harness for NVIDIA Nemotron 3 Nano 30B-A3B, tuned for TypeScript/JavaScript code agents"
readme = "README.md"
requires-python = ">=3.11,<3.13"
license = {text = "Apache-2.0"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["llm", "nemotron", "typescript", "fine-tuning", "qlora", "code-generation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    # Core ML
    "torch>=2.4.0",
    "transformers>=4.45.0",
    "accelerate>=0.34.0",
    "peft>=0.13.0",
    "bitsandbytes>=0.44.0",
    "datasets>=3.0.0",
    "safetensors>=0.4.0",

    # Quantization
    "auto-gptq>=0.7.0",

    # Data processing
    "tiktoken>=0.7.0",
    "sentencepiece>=0.2.0",
    "tokenizers>=0.20.0",

    # Utilities
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "tqdm>=4.66.0",
    "rich>=13.0.0",
    "click>=8.1.0",
    "httpx>=0.27.0",
    "aiohttp>=3.10.0",

    # Data deduplication
    "datasketch>=1.6.0",
    "xxhash>=3.4.0",

    # GitHub API
    "PyGithub>=2.4.0",

    # Logging and monitoring
    "wandb>=0.18.0",
    "tensorboard>=2.17.0",

    # Evaluation
    "evaluate>=0.4.0",
    "rouge-score>=0.1.2",
    "nltk>=3.9.0",

    # OpenAI-compatible client for vLLM
    "openai>=1.50.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "mypy>=1.11.0",
    "ruff>=0.6.0",
    "pre-commit>=3.8.0",
]

vllm = [
    "vllm>=0.6.0",
    "ray>=2.35.0",
]

nemo = [
    "nemo-toolkit[all]>=2.0.0",
]

[project.scripts]
nemotron-train = "src.train.dapt.train_hf_peft:main"
nemotron-eval = "src.eval.nemo_eval_runner:main"
collect-repos = "src.data.github.github_client:main"
build-corpus = "src.data.corpus.pack_jsonl:main"

[project.urls]
Homepage = "https://github.com/yourusername/nemotron-ts-posttrain-harness"
Documentation = "https://github.com/yourusername/nemotron-ts-posttrain-harness#readme"
Repository = "https://github.com/yourusername/nemotron-ts-posttrain-harness"
Issues = "https://github.com/yourusername/nemotron-ts-posttrain-harness/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "UP",   # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
asyncio_mode = "auto"

[tool.coverage.run]
source = ["src"]
omit = ["tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]
